패키지, 기능 세분화
