+++ feature/feed/src/main/java/com/dnd_9th_3_android/gooding/feed/MainLayout/NowFeedScreen.kt
@@ -4,19 +4,19 @@ import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.colorResource
 import com.dnd_9th_3_android.gooding.feature.feed.R
 
+// 메인 피드 - now
 @Composable
-fun RecommendScreen() {
+fun NowScreen() {
     Column(
         modifier = Modifier
             .fillMaxWidth()
             .fillMaxHeight()
-            .background(colorResource(id = R.color.type2))
+            .background(colorResource(id = R.color.type1))
     ) {
 
     }
